requirements.txt
solana-sdk==0.30.2
solders==0.18.1
customtkinter==5.2.1
pillow==10.2.0
base58==2.1.1 